+++ src/main/java/io/nagurea/smsupsdk/accountmanaging/account/retrieve/response/account/RetrieveAccountResultResponse.java
@@ -1,4 +1,4 @@
-package io.nagurea.smsupsdk.accountmanaging.account.retrieve.response;
+package io.nagurea.smsupsdk.accountmanaging.account.retrieve.response.account;
 
 import io.nagurea.smsupsdk.common.response.ResultResponse;
 import io.nagurea.smsupsdk.common.status.ResponseStatus;